+++ src/Bridges/Constraint/indicator_activate_on_zero.jl
@@ -21,7 +21,7 @@ struct IndicatorActiveOnFalseBridge{
     }
     indicator_cons_index::MOI.ConstraintIndex{
         F,
-        MOI.IndicatorSet{MOI.ACTIVATE_ON_ONE,S},
+        MOI.Indicator{MOI.ACTIVATE_ON_ONE,S},
     }
 end
 
@@ -34,7 +34,7 @@ function bridge_constraint(
     S<:MOI.AbstractScalarSet,
     T<:Real,
     F,
-    IS<:MOI.IndicatorSet{MOI.ACTIVATE_ON_ZERO,S},
+    IS<:MOI.Indicator{MOI.ACTIVATE_ON_ZERO,S},
 }
     f_scalars = MOIU.eachscalar(f)
     z2, zo_cons = MOI.add_constrained_variable(model, MOI.ZeroOne())
@@ -45,7 +45,7 @@ function bridge_constraint(
     ci = MOI.add_constraint(
         model,
         f2,
-        MOI.IndicatorSet{MOI.ACTIVATE_ON_ONE}(s.set),
+        MOI.Indicator{MOI.ACTIVATE_ON_ONE}(s.set),
     )
     return IndicatorActiveOnFalseBridge{T,F,S}(z2, zo_cons, dcons, ci)
 end
@@ -53,7 +53,7 @@ end
 function MOI.supports_constraint(
     ::Type{<:IndicatorActiveOnFalseBridge{T}},
     ::Type{<:MOI.VectorAffineFunction},
-    ::Type{<:MOI.IndicatorSet{MOI.ACTIVATE_ON_ZERO}},
+    ::Type{<:MOI.Indicator{MOI.ACTIVATE_ON_ZERO}},
 ) where {T}
     return true
 end
@@ -69,14 +69,14 @@ function MOIB.added_constraint_types(
 ) where {T,F,S}
     return [
         (MOI.ScalarAffineFunction{T}, MOI.EqualTo{T}),
-        (F, MOI.IndicatorSet{MOI.ACTIVATE_ON_ONE,S}),
+        (F, MOI.Indicator{MOI.ACTIVATE_ON_ONE,S}),
     ]
 end
 
 function concrete_bridge_type(
     ::Type{<:IndicatorActiveOnFalseBridge{T}},
     ::Type{F},
-    ::Type{MOI.IndicatorSet{MOI.ACTIVATE_ON_ZERO,S}},
+    ::Type{MOI.Indicator{MOI.ACTIVATE_ON_ZERO,S}},
 ) where {T,F<:MOI.VectorAffineFunction,S<:MOI.AbstractScalarSet}
     return IndicatorActiveOnFalseBridge{T,F,S}
 end
@@ -84,7 +84,7 @@ end
 function concrete_bridge_type(
     ::Type{<:IndicatorActiveOnFalseBridge},
     ::Type{F},
-    ::Type{MOI.IndicatorSet{MOI.ACTIVATE_ON_ZERO,S}},
+    ::Type{MOI.Indicator{MOI.ACTIVATE_ON_ZERO,S}},
 ) where {F<:MOI.VectorAffineFunction,S<:MOI.AbstractScalarSet}
     return IndicatorActiveOnFalseBridge{Float64,F,S}
 end