+++ NEWS.md
@@ -18,7 +18,8 @@ Breaking renames and type changes:
 - `Constraint.SOCRBridge` has been renamed to `Constraint.SOCtoRSOCBridge`
 - `RawParameter` has been renamed to `RawOptimizerAttribute` and now takes
   `String` instead of `Any` as the only argument
-- `Test.TestConfig` has been renamed to `Test.Config`
+- `Test.TestConfig` has been renamed to `Test.Config`    
+- `IndicatorSet` has been renamed to `Indicator`
 
 Breaking behavior changes: